+++ api/src/api/page/controllers/page.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * page controller
+ */
+
+const { createCoreController } = require('@strapi/strapi').factories;
+
+module.exports = createCoreController('api::page.page');
