+++ api/src/api/page/routes/page.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * page router
+ */
+
+const { createCoreRouter } = require('@strapi/strapi').factories;
+
+module.exports = createCoreRouter('api::page.page');
