+++ api/src/api/page/services/page.js
@@ -1,0 +1,9 @@
+'use strict';
+
+/**
+ * page service
+ */
+
+const { createCoreService } = require('@strapi/strapi').factories;
+
+module.exports = createCoreService('api::page.page');
